+++ services/service-neis/src/main/kotlin/com/b1nd/dodamdodam/neis/infrastructure/comcigan/data/ParsedSchedule.kt
@@ -1,0 +1,12 @@
+package com.b1nd.dodamdodam.neis.infrastructure.comcigan.data
+
+import java.time.LocalDate
+
+data class ParsedSchedule(
+    val date: LocalDate,
+    val grade: Int,
+    val room: Int,
+    val period: Int,
+    val subject: String,
+    val teacher: String,
+)
